+++ src/Exception/InvalidConfigurationException.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace CrazyGoat\ScanMePHP\Exception;
+
+use Exception;
+
+class InvalidConfigurationException extends Exception
+{
+    public static function invalidModuleSize(int $size): self
+    {
+        return new self(sprintf(
+            'Module size must be greater than 0, got %d',
+            $size
+        ));
+    }
+}